requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
asciichartpy==1.5.25
attrs==24.2.0
boolean.py==4.0
bottle==0.13.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
contourpy==1.3.0
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.1.0
grpcio==1.66.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
huggingface-hub==0.24.6
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
optuna==4.0.0
optuna-dashboard==0.16.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
pluggy==1.5.0
protobuf==5.28.0
psutil==6.0.0
Pympler==1.1
pyparsing==3.1.4
pyright==1.1.385
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
safetensors==0.4.4
sb3_contrib==2.3.0
scikit-learn==1.5.1
scipy==1.14.1
setuptools==72.1.0
shodan==1.31.0
six==1.16.0
SQLAlchemy==2.0.35
stable_baselines3==2.3.2
sympy==1.13.2
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.19.1
torch==2.4.0
torch_geometric==2.5.3
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ufal.chu_liu_edmonds==1.0.3
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.43.0
XlsxWriter==3.2.0
yarl==1.9.8
